back/pyproject.toml
[tool.poetry]
name = "colosseum-backend"
version = "0.1.0"
description = "Water management hackathon backend for Solana Cypherpunk"
authors = ["Team Colosseum"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.0"
solana = "^0.34.0"
solders = "^0.21.0"
httpx = "^0.27.0"
python-dateutil = "^2.9.0"
pillow = "^11.0.0"
base58 = "^2.1.1"
anchorpy = "^0.20.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
ruff = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
